Address PR #278 follow-up review: scope staging token, align cancel-in-progress, dedup spec helper
Six items from claude[bot]'s 2026-04-29 review on top of bfd0f3b.

cpflow-deploy-staging.yml
- Drop workflow-level CPLN_TOKEN entry. Authentication now flows through the
  cpln profile written by cpflow-setup-environment, matching the pattern
  already in place for cpflow-deploy-review-app.yml and cpflow-delete-review-app.yml.
- Switch concurrency.cancel-in-progress from true to false to match the review-app
  and delete workflows. A cancelled cpflow deploy-image mid-rollout can leave the
  staging GVC in a partially-deployed state; let in-flight deploys finish.

cpflow-delete-control-plane-app/delete-app.sh
- Add a TODO noting the string-based cpflow exists detection is a short-term
  workaround until cpflow exposes a structured signal (distinct exit code or
  --json output) for "not found" vs. API/auth errors.

cpflow-promote-staging-to-production.yml
- Expand the HEALTH_CHECK_ACCEPTED_STATUSES comment to explain why the default
  accepts 301/302: curl runs without -L, and the redirect is the most common
  signal that an auth-gated root path is reachable. Override guidance updated.

github_flow_readiness_service.rb
- Expand the fetch_with_cache comment to document why the mutex is released before
  yield (avoid blocking other workers during HTTP I/O), why parallel cache misses
  are bounded (idempotent fetches, per-name keys), and why the duplication only
  occurs on the first parallel sweep.

spec/support/directory_helpers.rb (new)
- Extract the inside_dir helper that had been copy-pasted into four spec files
  (generate_spec, generate_github_actions_spec, github_flow_readiness_spec,
  ai_github_flow_prompt_spec). Included via spec_helper alongside the other
  shared helpers.

Verification
- bundle exec rspec on the five spec files covering this PR (generate_spec,
  generate_github_actions_spec, github_flow_readiness_spec,
  ai_github_flow_prompt_spec, github_flow_readiness_service_spec) — 41 examples,
  0 failures, same count as bfd0f3b.
- bundle exec rubocop on all modified Ruby files — clean.
