+++ spec/command/ai_github_flow_prompt_spec.rb
@@ -5,14 +5,6 @@ require "pathname"
 require "tmpdir"
 
 describe Command::AiGithubFlowPrompt, :enable_validations, :without_config_file do
-  def inside_dir(path)
-    original_working_dir = Dir.pwd
-    Dir.chdir(path)
-    yield if block_given?
-  ensure
-    Dir.chdir(original_working_dir)
-  end
-
   let(:playground) { Pathname.new(Dir.mktmpdir("cpflow-ai-github-flow-prompt")) }
 
   after do
