+++ spec/spec_helper.rb
@@ -130,6 +130,7 @@ RSpec.configure do |config|
 
   config.include CommandHelpers
   config.include DateTimeHelpers
+  config.include DirectoryHelpers
   config.include StubENV
   config.include VerifiedDouble
